Updated monitor and first try at a frontend
Monitor.py is updated with some more user customization, a better
console output, and averaging to remove some noise. Frontend.py is a
threaded wxPython application that simulates aquiring data from the
IC/5. Unfortunately, it doesn't work hardly at all at the moment.

Frontend.py needs autoscaling of the plot axes, and to fix the labels
for rate and thickness. That's a minimum working application. From there
it can be expanded.