Rewrote the UI completely, added composition estimate
The UI has been completely rewritten to be less flexible, but better
looking. Settings are now contained in a single frame, with three tabs
for the different sections. The deposition settings tab is a first step
towards multi-layer support. The rates graph has been broken out to a
separate frame and a composition plot has been added.

Primary goals now are redoing the deposition clock to update independent
of the backend, add more controls for the graphs, and to add UI
flexibility back in.